backend/requirements.txt
fastapi==0.115.6
uvicorn[standard]==0.34.0
sqlalchemy==2.0.36
pydantic==2.10.4
python-jose[cryptography]==3.3.0
passlib==1.7.4
python-multipart==0.0.20
alembic==1.14.1
pytest==8.3.4
httpx==0.28.1
